pyproject.toml
[tool.poetry]
name = "betfairstreamer"
version = "0.2.1"
description = ""
authors = ["Jonatan Almen <almen.jonatan@gmail.com>"]
packages = [
    { include = "betfairstreamer" }
]
[tool.poetry.dependencies]
python = "^3.8.0"
betfairlightweight = "^2.0.1"
orjson = "^2.5.0"
numpy = "^1.18.1"
pyzmq = "^18.1.1"
ciso8601 = "^2.1.3"

[tool.poetry.dev-dependencies]
pytest = "^5.3.5"
hypothesis = "^5.5.4"
pytest-mypy = "^0.4.2"
pytest-cov = "^2.8.1"
pytest-watch = "^4.2.0"
